+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.1-alpha.1
+
+* Limits the scope of application of the method while the filter is applied.
+  - remove, removeWhere, retainWhere, clear.
+* Add method.
+  - removeFromOriginal, removeWhereFromOriginal, retainWhereFromOriginal, clearFromOriginal.
+
 ## 0.0.1-alpha.0
 
-- Initial version, created by Stagehand
+* Initial version, created by Stagehand